New version of the scrape.py
The tokenizing function is deliberately "lenient" because im processing the text to be used for Doc2Vec for sentiment analysis. I didnt' include the code for that because Im still working on it.

This scraper deals with the "load more" buttons on websites and right now it is only set up for one website which is linked in the code.  i cant really explain it through comments so maybe I can walk yall through it during next* meeting.  Right now it is only for cointelegraph

But feel free to mess around with the code and see what else you can build from it if it helps. Let me know if you find any bugs.  The downloading article text bit takes a bit of time because I think the “download hasn’t started” error was happening due to the website blocking the requests (since it was happening too frequently)